NO-JIRA: Fixed compile error on RHEL5: double doesn't convert to int64_t
git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk/qpid@1360282 13f79535-47bb-0310-9956-ffa450edef68

+++ cpp/src/qpid/ha/Primary.cpp
@@ -99,7 +99,7 @@ Primary::Primary(HaBroker& hb, const BrokerInfo::Set& expect) :
             if (!backup->isReady()) expectedBackups.insert(backup);
         }
         // Set timeout for expected brokers to connect and become ready.
-        sys::Duration timeout(hb.getSettings().backupTimeout*sys::TIME_SEC);
+        sys::Duration timeout(int64_t(hb.getSettings().backupTimeout*sys::TIME_SEC));
         sys::AbsTime deadline(sys::now(), timeout);
         timerTask = new ExpectedBackupTimerTask(*this, deadline);
         hb.getBroker().getTimer().add(timerTask);
